+++ packages/orq-rc/src/models/operations/retrieveagentrequest.ts
@@ -1154,7 +1154,7 @@ export type RetrieveAgentRequestResponseBody = {
   type: RetrieveAgentRequestType;
   role: string;
   description: string;
-  systemPrompt?: string | undefined;
+  systemPrompt?: string | null | undefined;
   instructions: string;
   settings?: RetrieveAgentRequestSettings | undefined;
   model: RetrieveAgentRequestModel;
@@ -2811,7 +2811,7 @@ export const RetrieveAgentRequestResponseBody$inboundSchema: z.ZodType<
   type: RetrieveAgentRequestType$inboundSchema.default("internal"),
   role: z.string(),
   description: z.string(),
-  system_prompt: z.string().optional(),
+  system_prompt: z.nullable(z.string()).optional(),
   instructions: z.string(),
   settings: z.lazy(() => RetrieveAgentRequestSettings$inboundSchema).optional(),
   model: z.lazy(() => RetrieveAgentRequestModel$inboundSchema),