+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-03T15:25:58.138Z",
+    "2026-04-06T05:39:15.293Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {